chore(data): removing actions folder and opting for data fodler instead
